8052 Electronic Organ
This is a simple Embedded Systems project. The 8052-based
microcontroller was interfaced with a 4x4 keypad, an LCD and a speaker
to produce 16 different frequencies. The LCD would display the current
note and the frequency. The objective of this project was to have a
thorough understanding of interrupt based calculations and the
interfacing of the microcontroller with LCD and keyboard.